add symlink during scaffolding
